requirements.txt
django
djangorestframework
dj-rest-auth
django-allauth
django-rql
django-auto-prefetching
